refactor: rename models and update imports for blacklisted tokens and face swap
